WMR-229 Throw an error when using node built-ins
WMR does not polyfill node core modules to prompt users to
double-check their use in the browser and reduce bundle size.

This is a small change to add a thrown error message when this occurs
in production, in development a warning is given instead.
this informs the user which dependency requires the module and where in their code.