Prerendering: Remove incomplete fake document
Prerendering of meta tags and other elements in the documents head
previously worked by relying on a custom implementation of the global
`document` object. We only did add a tiny subset of the methods that are
available in the browser, but doing so breaks nearly all server
detection code in third party libs. Checked various libraries and the most
common snippet to do that is a variation of:

```js
const isServer = typeof document === "undefined";
```

Because we did patch in a global document object, this value is always
false, even during SSR.

Another problem with that approach in genral is that we'll always have
a biased subset of the API that may work with our code, but is not
guaranteed to work with third-party libraries. This is more complicated
because to check if a meta tag is already set libraries use the
`document.querySelector` api. And I feel like providing that bloats up
WMR unnecessarily and only leads us down the rabbit hole of adding more
and more DOM apis.

Instead a much more reliable approach is to ditch the whole idea of
supplying a fake document in the first place and instead rely on
consumers supplying the head value