Fix process.env import skewing env inlining offsets
This fixes #818. The issue was happening because `code` is passed to MagicString, which assumes all offsets are relative to that string. However, `code` was being modified to prepend an import statement before it gets scanned by the regex, which caused all offsets for `.overwrite()` to be relative to the modified `code` instead of the one that was passed to MagicString.